'datasette --get' option, refs #926
